refactor(receipts): extract validateAmexLinesForSignoff shared helper
Item 11: Move the AMEX line validation loop (match status, category,
receipt confirmation, attendee, business trip checks) into a single
pure function in reconciliation-signoff.ts. Both the finalize route
and month-closing now delegate to this helper, eliminating duplicate
validation logic.
