+++ app/api/receipts/reconcile/finalize/route.ts
@@ -12,9 +12,8 @@ import {
   listReceiptRecords,
 } from "@/lib/receipts/db";
 import { hashCsvContent } from "@/lib/receipts/export";
-import { buildReconciliationManifestCsv } from "@/lib/receipts/reconciliation-signoff";
+import { buildReconciliationManifestCsv, validateAmexLinesForSignoff } from "@/lib/receipts/reconciliation-signoff";
 import { archiveManifest, deleteArchiveObject } from "@/lib/receipts/storage";
-import { requiresAttendees } from "@/lib/receipts/categories";
 
 export async function POST(request: Request) {
   try {
@@ -49,7 +48,6 @@ export async function POST(request: Request) {
     const activeArtifact = await getAmexArtifactByMonth(month);
 
     // Validate: all lines must be resolved
-    const blockers: string[] = [];
     const amexAttendees = await listAmexLineAttendeeNamesByMonth(month);
     const receipts = await listReceiptRecords({ paymentPath: "AMEX", limit: 200 });
     const receiptAttendeeMap = new Map<string, string[]>();
@@ -63,42 +61,7 @@ export async function POST(request: Request) {
       if (entry) receiptAttendeeMap.set(entry[0], entry[1]);
     }
 
-    for (const line of amexLines) {
-      const label = `${line.transaction_date} ${line.merchant}`;
-
-      if (line.match_status === "unmatched" || line.match_status === "matched") {
-        blockers.push(`AMEX ${label}: unresolved match status (${line.match_status})`);
-      }
-      if (!line.expense_category_code) {
-        blockers.push(`AMEX ${label}: missing expense category`);
-      }
-      if (
-        line.receipt_status === "matched" &&
-        (!line.matched_receipt_id || line.match_status !== "confirmed")
-      ) {
-        blockers.push(`AMEX ${label}: matched receipt is not confirmed`);
-      }
-      if (
-        line.receipt_status === "missing_receipt" ||
-        ((line.receipt_status === "no_receipt_required" ||
-          line.receipt_status === "receipt_not_available") &&
-          !line.receipt_missing_reason)
-      ) {
-        blockers.push(`AMEX ${label}: missing receipt requires a reason`);
-      }
-      if (requiresAttendees(line.expense_category_code)) {
-        const linkedReceiptAttendees = line.matched_receipt_id
-          ? receiptAttendeeMap.get(line.matched_receipt_id) ?? []
-          : [];
-        const directAmexAttendees = amexAttendees[line.id] ?? [];
-        if (linkedReceiptAttendees.length === 0 && directAmexAttendees.length === 0) {
-          blockers.push(`AMEX ${label}: requires attendees`);
-        }
-      }
-      if (line.business_trip_status === "candidate") {
-        blockers.push(`AMEX ${label}: unresolved business trip candidate`);
-      }
-    }
+    const blockers = validateAmexLinesForSignoff(amexLines, amexAttendees, receiptAttendeeMap);
 
     if (blockers.length > 0) {
       return NextResponse.json(
